refactor: Order tests has been refactored
