+++ src/test/java/br/com/fiap/tech_challenge/core/domain/usecases/order/impl/CreateOrderUseCaseImplTest.java
@@ -4,7 +4,7 @@ import br.com.fiap.tech_challenge.core.domain.exceptions.DoesNotExistException;
 import br.com.fiap.tech_challenge.core.domain.models.Customer;
 import br.com.fiap.tech_challenge.core.domain.models.enums.ProductCategoryEnum;
 import br.com.fiap.tech_challenge.core.domain.models.enums.ProductStatusEnum;
-import br.com.fiap.tech_challenge.core.domain.models.order.Order;
+import br.com.fiap.tech_challenge.core.domain.models.Order;
 import br.com.fiap.tech_challenge.core.domain.models.product.Product;
 import br.com.fiap.tech_challenge.core.domain.ports.CustomerPersistence;
 import br.com.fiap.tech_challenge.core.domain.ports.OrderPersistence;
